requirements.txt
aniso8601==9.0.1
asgiref==3.4.1
bcrypt==3.2.2
cffi==1.15.1
click==8.1.3
colorama==0.4.5
distlib==0.3.5
Django==3.2.7
filelock==3.7.1
Flask==2.2.2
Flask-Bcrypt==1.0.1
Flask-Login==0.6.2
Flask-RESTful==0.3.9
gunicorn==20.1.0
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonify==0.5
MarkupSafe==2.1.1
mysql-connector-python==8.0.29
platformdirs==2.5.2
protobuf==4.21.9
pycparser==2.21
pytz==2021.1
six==1.16.0
sqlparse==0.4.2
virtualenv==20.15.1
Werkzeug==2.2.2
zipp==3.8.1
